hack/translate/pyproject.toml
[project]
name = "translate"
version = "0.0.0"
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.11.7",
    "python-liquid>=2.1.0",
    "pyyaml>=6.0.2",
    "tomlkit>=0.13.3",
    "types-pyyaml>=6.0.12.20250822",
]
